% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich_utils.R
\name{rich_aggregate}
\alias{rich_aggregate}
\title{aggregate enriched data}
\usage{
rich_aggregate(path = svDialogs::dlg_dir()$res)
}
\arguments{
\item{path}{The folder to look into, default is `svDialogs::dlg_dir()$res`}
}
\value{
A dataframe of enriched results.
}
\description{
The function looks recursively inside a folder and extract and bind together enrichment data
generated by \href{https://www.gsea-msigdb.org/gsea/index.jsp}{gsea}.
}
