% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich_utils.R
\name{rich_find}
\alias{rich_find}
\title{find files}
\usage{
rich_find(path = path, reg_expr)
}
\arguments{
\item{path}{The folder to look into}

\item{reg_expr}{Regexp to look}
}
\value{
A dataframe
}
\description{
The function looks recursively inside a folder and find specific files. Used internally.
}
