% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richer_graphs.R
\name{rich_pheatmap}
\alias{rich_pheatmap}
\title{rich_pheatmap}
\usage{
rich_pheatmap(
  dat,
  fdr_threshold,
  gs,
  value = "n_logp_sign",
  col = c("#053061", "#2166AC", "#4393C3", "#92C5DE", "#D1E5F0", "#F7F7F7", "#FDDBC7",
    "#F4A582", "#D6604D", "#B2182B", "#67001F"),
  ...
)
}
\arguments{
\item{dat}{Dataframe as returned by script.}

\item{fdr_threshold}{Minimum `FDR.q.val` to keep.}

\item{gs}{Gene Set.}

\item{value}{Column from which to take the value for pivit wider. One of c("fdr_q_val", "nes", "nom_p_val", "n_logp_sign").
Defaut is "n_logp_sign".}

\item{col}{Palette of heatmap. Default is `rev(RColorBrewer::brewer.pal(11, "RdBu"))`.}

\item{...}{\link[pheatmap]{pheatmap} arguments .}
}
\description{
Just a wrapper that runs \code{\link{rich_wider}} and produces a heatmap of the results using `{pheatmap}`.
}
