% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richgraphs.R
\name{rich_pheatmap_}
\alias{rich_pheatmap_}
\title{czmo_pheatmap_gsea_}
\usage{
rich_pheatmap_(dat, gs, p_th, res, width, height, save_path, ...)
}
\arguments{
\item{dat}{Dataframe}

\item{gs}{The gs to plot}

\item{p_th}{Threshold for p value filtering}

\item{res}{Resolution in dpi}

\item{width}{Width of the graph}

\item{height}{Height of the graph}

\item{save_path}{Where to save the data.}

\item{...}{Pheatmap arguments}
}
\description{
Given a dataframe with pathway, one gs, contrast, FDR.q.val and NES, it reshapes it, calculates
negative log10(fdr_q_val) * sign of NES and plots it
plots heatmaps. This is use internaly
}
