% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richgraphs.R
\name{rich_pheatmap}
\alias{rich_pheatmap}
\title{rich_pheatmap}
\usage{
rich_pheatmap(dat, p_th, res, width, height, save_path, ...)
}
\arguments{
\item{dat}{Dataframe.}

\item{p_th}{Threshold for p value filtering.}

\item{res}{Resolution in dpi.}

\item{width}{Width of the graph.}

\item{height}{Height of the graph.}

\item{save_path}{Where to save the data.}

\item{...}{pheatmap arguments.}
}
\description{
Given a dataframe with pathway, gs, contrast, FDR.q.val and NES, it reshapes it, calculates
\code{-log10(fdr_q_val) * sign(NES)} and plots a pheatmap. Values of FDR.q.val equal to 0 are
substituted with \code{10^10}. Plot is generate with \code{{pheatmap}} package and so function accepts those arguments too.
(Yes a method of of pheatmap would have been politer but this is just used internally in the lab)
}
