% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richer_graphs.R
\name{rich_pheatmap}
\alias{rich_pheatmap}
\title{rich_pheatmap}
\usage{
rich_pheatmap(dat, fdr_threshold, gs, value = "n_logp_sign", ...)
}
\arguments{
\item{dat}{Dataframe as returned by script.}

\item{fdr_threshold}{Minimum \code{FDR.q.val} to keep.}

\item{gs}{Gene Set.}

\item{col}{Palette of heatmap. Default is \code{rev(RColorBrewer::brewer.pal(11, "RdBu"))}.
#' @param ... \link[pheatmap]{pheatmap} arguments .}
}
\description{
Just a wrapper that runs \code{\link{rich_wider}} and produces a heatmap of the results using \code{{pheatmap}}.
}
