% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich_utils.R
\name{rich_wider}
\alias{rich_wider}
\title{rich_wider}
\usage{
rich_wider(dat, fdr_threshold, gs, value = "n_logp_sign")
}
\arguments{
\item{dat}{Dataframe as returned by \code{\link{rich_aggregate}}.}

\item{fdr_threshold}{Minimum `FDR.q.val` to keep.}

\item{gs}{Gene Set Database.}

\item{value}{Column from which to take the value for pivit wider. One of c("fdr_q_val", "nes", "nom_p_val", "n_logp_sign").
Defaut is "n_logp_sign".}
}
\description{
Given a dataframe generate by \code{\link{rich_aggregate}}, filters it to get a specific `gs`, it calculates
negative `log10(p.norm) * sign of NES`, filters out all `FDR.q.val >  fdr_threshold` and
returns a wider datagrame with `pathways` as rows, `contrasts` as
column and one of the columns as the `value`. When a combination of contrast-pathway is not presence a 0 is substituted
to NA when pivoting.
}
