% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich_utils.R
\name{rich_wider}
\alias{rich_wider}
\title{rich_wider}
\usage{
rich_wider(dat, fdr_threshold, gs, value = "n_logp_sign")
}
\arguments{
\item{dat}{Dataframe as returned by script.}

\item{fdr_threshold}{Minimum \code{FDR.q.val} to keep.}

\item{gs}{Gene Set.}

\item{value}{Column from which to take the value for pivit wider. One of c("fdr_q_val", "nes", "p_norm", "n_logp_sign").
Defaut is "n_logp_sign".}
}
\description{
Given a dataframe  with pathway, one gs, contrasts, FDR.q.val and NES, it calculates
negative \verb{log10(p.norm) * sign of NES}, filters out all \code{FDR.q.val >  fdr_threshold} and
returns a wider datagrame with \code{pathways} as rows, \code{contrasts} as
column and one of the columns as the value. When a combination of contrast-pathway is not presence a 0 is substituted
to NA.
}
